+ Move code to new ColumnWidthCalculator
+ Minor cleanups/re-arrangements.
