+++ svc/GridAutosizeService.js
@@ -6,22 +6,9 @@
  */
 
 import {HoistService} from '@xh/hoist/core';
-import {action} from '@xh/hoist/mobx';
-import {stripTags, throwIf} from '@xh/hoist/utils/js';
-import {
-    groupBy,
-    isFinite,
-    isFunction,
-    isNil,
-    map,
-    max,
-    min,
-    reduce,
-    sum,
-    compact,
-    sortBy,
-    isEmpty
-} from 'lodash';
+import {isFinite, map, sum, compact, sortBy, isEmpty} from 'lodash';
+import {runInAction} from '@xh/hoist/mobx';
+import {ColumnWidthCalculator} from './impl/ColumnWidthCalculator';
 
 /**
  * Calculates the column width required to display all values in a column.
@@ -39,9 +26,7 @@ import {
 @HoistService
 export class GridAutosizeService {
 
-    _canvasContext;
-    _headerEl;
-    _cellEl;
+    _columnWidthCalculator = new ColumnWidthCalculator();
 
     /**
      * Calculate and apply autosized column widths.
@@ -50,8 +35,7 @@ export class GridAutosizeService {
      * @param {String[]} colIds - array of columns in model to compute sizing for.
      * @param {GridAutosizeOptions} options - options to use for this autosize.
      */
-    @action
-    autosizeColumns(gridModel, colIds, options) {
+    async autosizeAsync(gridModel, colIds, options) {
         // 1) Remove any columns with element renderers
         colIds = colIds.filter(id => {
             const col = gridModel.getColumn(id);
@@ -60,20 +44,28 @@ export class GridAutosizeService {
         if (isEmpty(colIds)) return;
 
         // 2) Ensure order of passed colIds matches the current GridModel.columnState.
-        // This is to prevent inadvertently changing the column order when applying column state changes
+        // This is to prevent changing the column order when applying column state changes
         colIds = sortBy(colIds, id => gridModel.columnState.findIndex(col => col.colId === id));
+        runInAction(() => {
+            // 3) Shrink columns down to their required widths.
+            const requiredWidths = this.calcRequiredWidths(gridModel, colIds, options);
+            gridModel.applyColumnStateChanges(requiredWidths);
+            console.debug('Column widths autosized via GridAutosizeService', requiredWidths);
 
-        // 3) Shrink columns down to their required widths.
-        const requiredWidths = this.calcRequiredWidths(gridModel, colIds, options);
-        gridModel.applyColumnStateChanges(requiredWidths);
-        console.debug('Column widths autosized via GridAutosizeService', requiredWidths);
-
-        // 4) Grow columns to fill any remaining space, if enabled.
-        const fillWidths = this.calcFillWidths(gridModel, colIds, options);
-        gridModel.applyColumnStateChanges(fillWidths);
-        console.debug('Column widths filled via GridAutosizeService', fillWidths);
+            // 4) Grow columns to fill any remaining space, if enabled.
+            const {fillMode} = options;
+            if (fillMode && fillMode !== 'none') {
+                const fillWidths = this.calcFillWidths(gridModel, colIds, fillMode);
+                gridModel.applyColumnStateChanges(fillWidths);
+                console.debug('Column widths filled via GridAutosizeService', fillWidths);
+            }
+        });
     }
 
+
+    //------------------
+    // Implementation
+    //------------------
     /**
      * Calculates the required column widths for a GridModel. Returns an array of the
      * form [{colId, width}] suitable for consumption by GridModel.applyColumnStateChanges().
@@ -102,26 +94,26 @@ export class GridAutosizeService {
         }
 
         for (const colId of colIds) {
-            const width = this.autosizeColumn(gridModel, records, colId, options);
+            const width = this._columnWidthCalculator.calcWidth(gridModel, records, colId, options);
             if (isFinite(width)) ret.push({colId, width});
         }
 
         return ret;
     }
 
+
     /**
      * Calculate the increased size of columns to fill any remaining space. Returns an array of the
      * form [{colId, width}] suitable for consumption by GridModel.applyColumnStateChanges().
-     * Typically called via `GridModel.autosizeAsync()`.
      *
      * @param {GridModel} gridModel - GridModel to autosize.
      * @param {String[]} colIds - array of columns in model to compute sizing for.
-     * @param {GridAutosizeOptions} options - options to use for this autosize.
+     * @param {String} fillMode
      */
-    calcFillWidths(gridModel, colIds, options) {
-        const {fillMode} = options;
-        if (!fillMode || fillMode === 'none' || gridModel.getVisibleLeafColumns().some(it => it.flex)) return [];
-        throwIf(!['all', 'left', 'right'].includes(fillMode), `Unsupported value "${fillMode}" for fillMode.`);
+    calcFillWidths(gridModel, colIds, fillMode) {
+        if (gridModel.getVisibleLeafColumns().some(it => it.flex)) {
+            return [];
+        }
 
         // 1) Exclude pinned columns
         colIds = colIds.filter(id => !gridModel.getColumnPinned(id));
@@ -152,252 +144,7 @@ export class GridAutosizeService {
         }
     }
 
-    //------------------
-    // Implementation
-    //------------------
-    autosizeColumn(gridModel, records, colId, options) {
-        const column = gridModel.findColumn(gridModel.columns, colId),
-            {autosizeMinWidth, autosizeMaxWidth} = column;
-
-        let result = max([
-            this.calcHeaderMaxWidth(gridModel, column, options),
-            this.calcDataMaxWidth(gridModel, records, column, options)
-        ]);
-
-        result = max([result, autosizeMinWidth]);
-        result = min([result, autosizeMaxWidth]);
-        return result;
-    }
-
-    calcHeaderMaxWidth(gridModel, column, options) {
-        const {autosizeIncludeHeader, autosizeIncludeHeaderIcons} = column,
-            {bufferPx} = options;
-
-        if (!autosizeIncludeHeader) return null;
-
-        try {
-            this.setHeaderElActive(true);
-            this.setHeaderElSizingMode(gridModel.sizingMode);
-            return this.getHeaderWidth(gridModel, column, autosizeIncludeHeaderIcons, bufferPx);
-        } catch (e) {
-            console.warn(`Error calculating max header width for column "${column.colId}".`, e);
-        } finally {
-            this.setHeaderElActive(false);
-        }
-    }
-
-    calcDataMaxWidth(gridModel, records, column, options) {
-        try {
-            if (column.elementRenderer) return null;
-
-            const {store, sizingMode, treeMode} = gridModel;
-
-            this.setCellElActive(true);
-            this.setCellElSizingMode(sizingMode);
-
-            if (treeMode && column.isTreeColumn && store.allRootCount !== store.allCount) {
-                // For tree columns, we need to account for the indentation at the different depths.
-                // Here we group the records by tree depth and determine the max width at each depth.
-                const recordsByDepth = groupBy(records, record => record.ancestors.length),
-                    levelMaxes = map(recordsByDepth, (records, depth) => {
-                        return this.calcMaxWidth(gridModel, records, column, options, this.getIndentation(depth));
-                    });
-
-                return max(levelMaxes);
-            } else {
-                return this.calcMaxWidth(gridModel, records, column, options);
-            }
-        } catch (e) {
-            console.warn(`Error calculating max data width for column "${column.colId}".`, e);
-        } finally {
-            this.setCellElActive(false);
-        }
-    }
-
-    calcMaxWidth(gridModel, records, column, options, indentationPx = 0) {
-        const {field, getValueFn, renderer} = column,
-            {bufferPx, sampleCount} = options,
-            useRenderer = isFunction(renderer);
-
-        // 1) Get unique values
-        const values = new Set();
-        records.forEach(record => {
-            if (!record) return;
-            const rawValue = getValueFn({record, field, column, gridModel}),
-                value = useRenderer ? renderer(rawValue, {record, column, gridModel}) : rawValue;
-            values.add(value);
-        });
-
-        // 2) Use a canvas to estimate and sort by the pixel width of the string value.
-        // Strip html tags but include parentheses / units etc. for renderers that may return html,
-        const sortedValues = sortBy(Array.from(values), value => {
-            const width = isNil(value) ? 0 : this.getStringWidth(stripTags(value.toString()));
-            return width + indentationPx;
-        });
-
-        // 3) Extract the sample set of longest values for rendering and sizing
-        const longestValues = sortedValues.slice(Math.max(sortedValues.length - sampleCount, 0));
-
-        // 4) Render to a hidden cell to calculate the max displayed width
-        return reduce(longestValues, (currMax, value) => {
-            const width = this.getCellWidth(value, useRenderer) + indentationPx + bufferPx;
-            return Math.max(currMax, width);
-        }, 0);
-    }
-
-    //------------------
-    // Autosize header cell
-    //------------------
-    getHeaderWidth(gridModel, column, includeHeaderSortIcon, bufferPx) {
-        const {colId, headerName, agOptions} = column,
-            headerHtml = isFunction(headerName) ? headerName({column, gridModel}) : headerName,
-            showSort = column.sortable && (includeHeaderSortIcon || gridModel.sortBy.find(sorter => sorter.colId === colId)),
-            showMenu = agOptions?.suppressMenu === false;
-
-        // Render to a hidden header cell to calculate the max displayed width
-        const headerEl = this.getHeaderEl();
-        this.setHeaderElSortAndMenu(showSort, showMenu);
-        headerEl.firstChild.innerHTML = headerHtml;
-        return Math.ceil(headerEl.clientWidth) + bufferPx;
-    }
-
-    setHeaderElActive(active) {
-        const headerEl = this.getHeaderEl();
-        if (active) {
-            headerEl.classList.add('xh-grid-autosize-header--active');
-        } else {
-            headerEl.classList.remove('xh-grid-autosize-header--active');
-        }
-    }
-
-    setHeaderElSizingMode(sizingMode) {
-        const headerEl = this.getHeaderEl();
-        headerEl.classList.remove(
-            'xh-grid-autosize-header--large',
-            'xh-grid-autosize-header--standard',
-            'xh-grid-autosize-header--compact',
-            'xh-grid-autosize-header--tiny'
-        );
-        headerEl.classList.add(`xh-grid-autosize-header--${sizingMode}`);
-    }
-
-    setHeaderElSortAndMenu(sort, menu) {
-        const headerEl = this.getHeaderEl();
-        headerEl.classList.remove(
-            'xh-grid-autosize-header--sort',
-            'xh-grid-autosize-header--menu'
-        );
-        if (sort) headerEl.classList.add('xh-grid-autosize-header--sort');
-        if (menu) headerEl.classList.add('xh-grid-autosize-header--menu');
-    }
-
-    getHeaderEl() {
-        if (!this._headerEl) {
-            const headerEl = document.createElement('div');
-            headerEl.classList.add('xh-grid-autosize-header');
-            headerEl.appendChild(document.createElement('span'));
-
-            const sortIcon = document.createElement('div');
-            sortIcon.classList.add('xh-grid-header-sort-icon');
-            headerEl.appendChild(sortIcon);
-
-            const menuIcon = document.createElement('div');
-            menuIcon.classList.add('xh-grid-header-menu-icon');
-            headerEl.appendChild(menuIcon);
-
-            document.body.appendChild(headerEl);
-            this._headerEl = headerEl;
-        }
-        return this._headerEl;
-    }
-
-    //------------------
-    // Autosize cell
-    //------------------
-    getCellWidth(value, useRenderer) {
-        const cellEl = this.getCellEl();
-        if (useRenderer) {
-            cellEl.innerHTML = value;
-        } else if (cellEl.firstChild?.nodeType === 3) {
-            // If we're not rendering html and the cell's first child is already a TextNode,
-            // we can update it's data to avoid creating a new TextNode.
-            cellEl.firstChild.data = value;
-        } else {
-            cellEl.innerText = value;
-        }
-        return Math.ceil(cellEl.clientWidth);
-    }
-
-    setCellElActive(active) {
-        const cellEl = this.getCellEl();
-        if (active) {
-            cellEl.classList.add('xh-grid-autosize-cell--active');
-        } else {
-            cellEl.classList.remove('xh-grid-autosize-cell--active');
-        }
-    }
-
-    setCellElSizingMode(sizingMode) {
-        const cellEl = this.getCellEl();
-        cellEl.classList.remove(
-            'xh-grid-autosize-cell--large',
-            'xh-grid-autosize-cell--standard',
-            'xh-grid-autosize-cell--compact',
-            'xh-grid-autosize-cell--tiny'
-        );
-        cellEl.classList.add(`xh-grid-autosize-cell--${sizingMode}`);
-    }
-
-    getCellEl() {
-        if (!this._cellEl) {
-            const cellEl = document.createElement('div');
-            cellEl.classList.add('xh-grid-autosize-cell');
-            document.body.appendChild(cellEl);
-            this._cellEl = cellEl;
-        }
-        return this._cellEl;
-    }
-
-    //-----------------
-    // Indentation
-    //-----------------
-    getIndentation(depth) {
-        const cellEl = this.getCellEl(),
-            indentation = parseInt(window.getComputedStyle(cellEl).getPropertyValue('left'));
-        depth = parseInt(depth) + 1; // Add 1 to account for expand / collapse arrow.
-        return indentation * depth;
-    }
-
-    //------------------
-    // Canvas-based width estimation
-    //------------------
-    getStringWidth(string) {
-        const canvasContext = this.getCanvasContext();
-        return canvasContext.measureText(string).width;
-    }
-
-    getCanvasContext() {
-        if (!this._canvasContext) {
-            // Create hidden canvas
-            const canvasEl = document.createElement('canvas');
-            canvasEl.classList.add('xh-grid-autosize-canvas');
-            document.body.appendChild(canvasEl);
-
-            // Create context which uses grid fonts
-            const canvasContext = canvasEl.getContext('2d');
-            canvasContext.font = 'var(--xh-grid-font-size-px) var(--xh-grid-font-family)';
-            this._canvasContext = canvasContext;
-        }
-        return this._canvasContext;
-    }
-
-    //------------------
-    // Column width filling
-    //------------------
-
-    /**
-     * Divide the remaining space evenly amongst columns, while respecting their maxWidths.
-     */
+    // Divide the remaining space evenly amongst columns, while respecting their maxWidths.
     fillEvenly(columns, remaining) {
         const ret = {};
 
@@ -425,9 +172,7 @@ export class GridAutosizeService {
         });
     }
 
-    /**
-     * Divide the remaining space across columns in order, while respecting their maxWidths.
-     */
+    // Divide the remaining space across columns in order, while respecting their maxWidths.
     fillSequentially(columns, remaining) {
         const ret = [];
         for (const col of columns) {
@@ -445,35 +190,31 @@ export class GridAutosizeService {
     }
 
     getFillState(gridModel, colIds) {
-        const total = this.getTotalColumnWidth(gridModel),
-            columns = compact(colIds.map(colId => this.getColumnFillState(gridModel, colId)));
-
-        return {total, columns};
+        return {
+            total: this.getTotalColumnWidth(gridModel),
+            columns: compact(colIds.map(colId => this.getColumnFillState(gridModel, colId)))
+        };
     }
 
-    /**
-     * Returns an array of column state representations,
-     * suitable for use by the column fill algorithms.
-     */
+    // Returns an array of column state representations,
+    // suitable for use by the column fill algorithms.
     getColumnFillState(gridModel, colId) {
         const column = gridModel.getColumn(colId),
             colState = gridModel.getStateForColumn(colId);
 
         if (!column || !colState || colState.hidden) return null;
 
-        const {width} = colState,
-            maxWidth = column.autosizeMaxWidth;
-
-        return {colId, width, maxWidth};
+        return {
+            colId,
+            width: colState.width,
+            maxWidth: column.autosizeMaxWidth
+        };
     }
 
-    /**
-     * Returns the total combined width of visible columns.
-     * Note that this intentionally excludes pinned columns.
-     */
+    // Returns the total combined width of visible columns.
+    // Note that this intentionally excludes pinned columns.
     getTotalColumnWidth(gridModel) {
         const widths = gridModel.columnState.filter(it => !it.hidden && !it.pinned).map(it => it.width);
         return sum(widths);
     }
-
 }
